refactor: switch to application factory pattern
We now use Flask application factory pattern to be able to create
multiple instances of the app with different configurations.
This allows us to have different configurations for testing:
e.g. we can now create a separate flask instance with a temporary
database that we can create and destroy with each test.

This major update includes updates to the tests and all blueprints.